README
Target Codevita 2020 : Contest 20

July 30 2020 11:00 AM to July 31 2020 11:00 AM

Description:

These are solutions to the coding problems of the second challenge. Problems are taken from Hackerrank and TCS Codevita challenge. All the codes are compiled at Hackerrank editor and coded in the form of functions. Most of the solutions are built in an easy manner, so that new learners can also understand the code. Codes are well commented.

The link to the problems are given below.

Happy Coding.

https://www.hackerrank.com/challenges/hackerland-radio-transmitters

https://www.hackerrank.com/challenges/strange-grid

https://www.hackerrank.com/challenges/identify-smith-numbers