Hackerland-radio-transmitters-statement
Problem Statement:
Hackerland is a one-dimensional city with houses aligned at integral locations along a road. The Mayor wants to install radio transmitters on the roofs of the city's houses. Each transmitter has a fixed range meaning it can transmit a signal to all houses within that number of units distance away.

Given a map of Hackerland and the transmission range, determine the minimum number of transmitters so that every house is within range of at least one transmitter. Each transmitter must be installed on top of an existing house.

For example, assume houses are located at x=[1,2,3,5,9] and the transmission range k=1. 3 antennae at houses 2 and 5 and 9 would provide complete coverage. There is no house at location 7 to cover both 5 and 9. Ranges of coverage, are [1,2,3], [5], and [9].

Function Description:
Complete the hackerlandRadioTransmitters function in the editor below. It must return an integer that denotes the minimum number of transmitters to install.

hackerlandRadioTransmitters has the following parameter(s):
    x: integer array that denotes the locations of houses
    k: an integer that denotes the effective range of a transmitter


For more details, refer
https://www.hackerrank.com/challenges/hackerland-radio-transmitters